Created 2 objects to replace functions that were on the Plugin class: AuthorizationServiceProvider and AuthenticationServiceProvider. Both of these class implement the respective interfaces previously on the Plugin class. Both classes are now passed MiddlewareQueueLoader and type hints that were to the concretion "Plugin" are now the abstraction interfaces.
