+++ src/Plugin.php
@@ -13,17 +13,16 @@ declare(strict_types=1);
 
 namespace CakeDC\Users;
 
-use Authentication\AuthenticationServiceInterface;
-use Authentication\AuthenticationServiceProviderInterface;
-use Authorization\AuthorizationServiceInterface;
-use Authorization\AuthorizationServiceProviderInterface;
 use Cake\Core\BasePlugin;
-use Cake\Core\Configure;
 use Cake\Http\MiddlewareQueue;
-use Psr\Http\Message\ServerRequestInterface;
+use CakeDC\Users\Provider\AuthenticationServiceProvider;
+use CakeDC\Users\Provider\AuthorizationServiceProvider;
+use CakeDC\Users\Provider\ServiceProviderLoaderTrait;
 
-class Plugin extends BasePlugin implements AuthenticationServiceProviderInterface, AuthorizationServiceProviderInterface
+class Plugin extends BasePlugin
 {
+    use ServiceProviderLoaderTrait;
+
     /**
      * Plugin name.
      *
@@ -44,65 +43,16 @@ class Plugin extends BasePlugin implements AuthenticationServiceProviderInterfac
     public const EVENT_AFTER_RESEND_TOKEN_VALIDATION = 'Users.Global.afterResendTokenValidation';
 
     /**
-     * Returns an authentication service instance.
-     *
-     * @param \Psr\Http\Message\ServerRequestInterface $request Request
-     * @return \Authentication\AuthenticationServiceInterface
-     */
-    public function getAuthenticationService(ServerRequestInterface $request): AuthenticationServiceInterface
-    {
-        $key = 'Auth.Authentication.serviceLoader';
-
-        return $this->loadService($request, $key);
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function getAuthorizationService(ServerRequestInterface $request): AuthorizationServiceInterface
-    {
-        $key = 'Auth.Authorization.serviceLoader';
-
-        return $this->loadService($request, $key);
-    }
-
-    /**
      * @inheritDoc
      */
     public function middleware(MiddlewareQueue $middlewareQueue): MiddlewareQueue
     {
         $loader = $this->getLoader('Users.middlewareQueueLoader');
 
-        return $loader($middlewareQueue, $this);
-    }
-
-    /**
-     * Load a service defined in configuration $loaderKey
-     *
-     * @param \Psr\Http\Message\ServerRequestInterface $request The request.
-     * @param string $loaderKey service loader key
-     * @return mixed
-     */
-    protected function loadService(ServerRequestInterface $request, $loaderKey)
-    {
-        $serviceLoader = $this->getLoader($loaderKey);
-
-        return $serviceLoader($request);
-    }
-
-    /**
-     * Get the loader callable
-     *
-     * @param string $loaderKey loader configuration key
-     * @return callable
-     */
-    protected function getLoader($loaderKey)
-    {
-        $serviceLoader = Configure::read($loaderKey);
-        if (is_string($serviceLoader)) {
-            $serviceLoader = new $serviceLoader();
-        }
-
-        return $serviceLoader;
+        return $loader(
+            $middlewareQueue,
+            new AuthenticationServiceProvider(),
+            new AuthorizationServiceProvider()
+        );
     }
 }
